MB-8295: Parametrise Set View ETS tables
ETS tables a global. With this commit it's possible
to start couch_set_view with a paramter ('prod' or
'dev') and the ETS tables will be set accordingly.
They won't share the tables, but have their own ones.

This commit is just a refactoring, so it doesn't
really change the functionality or the external APIs.

This commit needs to be merged together with:
http://review.couchbase.org/26383

Change-Id: I4ed8093b402c041ba3f0e9842887b2547fd4cdc7
Reviewed-on: http://review.couchbase.org/26344
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>