MB-8689 Add couch_set_view:does_group_misses_deletes/2
This function checks if a view group missed processing of any
deleted documents that got purged by database compaction. This
performs an estimation only, as a 'true' response really means
'maybe'. This is because after a database compaction with purging
enabled, it's impossible for a view group to know exactly which
deleted documents were purged.

Change-Id: Ib8f661a820a52203be8d66ca0da647048749acdd
Reviewed-on: http://review.couchbase.org/27864
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>