Don't log missing document stack traces
Several times in the past, the SDKs (.Net at least) had the
habit of periodically querying /bucketname/_design for no reason,
which causes a document not found error to be thrown and a stack
tracing spawning several lines to be logged, example:

[couchdb:error,2013-08-28T12:59:16.540,ns_1@84.41.130.187:<0.16168.505>:couch_log:error:42]Uncaught error in HTTP request: {throw,{not_found,missing}}

Stacktrace: [{capi_frontend,couch_doc_open,3},
             {couch_httpd_db,db_doc_req,3},
             {request_throttler,do_request,3},
             {couch_httpd,handle_request,6},
             {mochiweb_http,headers,5},
             {proc_lib,init_p_do_apply,3}]
[couchdb:info,2013-08-28T12:59:16.548,ns_1@84.41.130.187:<0.16168.505>:couch_log:info:39]127.0.0.1 - - GET /default/_design 404

This change just eliminates the stack trace and explicit error
message. The URL request log line with a status of 404 is enough
to identify this, and much less verbose.

Change-Id: I0a13bad5137946cdf7c9033816e18b2565a3b599
Reviewed-on: http://review.couchbase.org/28679
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>