Don't report queue_shutdown as an error
It's not an error, but something expected when there's
a limit for a query and the view has more rows than that
limit. Therefore don't report the queue shutdown as an
error with a stack trace, such as recently observed in
MB-8957:

[couchdb:error,2013-08-27T17:45:47.323,ns_1@10.154.184.218:<0.10610.208>:couch_log:error:42]Caught unexpected error while serving view query sasl/_design/d2: queue_shutdown
[{couch_view_merger_queue,queue,2},
 {couch_view_merger,'-make_map_set_fold_fun/4-fun-0-',4},
 {couch_set_view,map_view_merge_callback,2},
 {couch_index_merger,collect_rows,4},
 {couch_index_merger,handle_skip,1},
 {couch_index_merger,'-merge_indexes_no_acc/2-fun-0-',3},
 {couch_view_merger,merge_map_views,1},
 {couch_index_merger,do_query_index,4}]

This was not always logged as it depends on specific timings
of process shutdowns, which happened to be observed some often
in a Windows cluster.

Change-Id: I5facc9ed823a0a3aa030c6d8a9c74303135e4f6f
Reviewed-on: http://review.couchbase.org/28680
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>