MB-9132: Update to newest etap
Apache CouchDB has a more recent version of etap that
contains the fixes that were previously applied to
the Couchbase fork.

The new version is a single module and also compiles well
with Erlang R16.

Having the same single module version as Apache CouchDB
enables GeoCouch to run the same etap tests on both,
Couchbase and Apache CouchDB.

Change-Id: I27cb726240eb91c03fb25032e2ed197494a47ffd
Reviewed-on: http://review.couchbase.org/29025
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>