Add V8 include path to CPPFLAGS
Without this change the header check would succeed only if another v8.h
is located in a standard header location (/usr/include, etc) or if the
environment variable C_INCLUDE_PATH was defined and pointing to the custom
location (which used to be my case). This ensures we can build when the
only v8.h is located in a custom path supplied by --with-v8-include.

Change-Id: I35acfb34d37fd5aea2b8cb119a22edb9d9e98d87
Reviewed-on: http://review.couchbase.org/30134
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>