MB-8454 Fix initial index build progress stats
During initial index build, calculation of total number of changes is
inaccurate since it is assumed to be equal to the sum of current seqs
for vbuckets. This results in displaying progress stats having total_changes
much more than the actual number of changes to be processed.

For initial build case, total changes should be estimated by summation
of non-deleted docs count from each of the vbuckets belonging to a bucket.

Change-Id: Iea4e2a6ebcd614d973ac5551e46c565e3be46f7c
Reviewed-on: http://review.couchbase.org/30202
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>