MB-9085: Don't hard-code UPR port
Instead of hard-coding the UPR port, use a configuration
setting for it. This makes it also possible to start the
fake UPR server on a random port (assigned by the
operating system), when the value is set to "0".

Hence it's now possible to run the set_view_tests in
parallel again.

Change-Id: I6c319dbb6fcac8b4f20f38ce8a44c6abed24658d
Reviewed-on: http://review.couchbase.org/29404
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>