MB-9776: Throw error on queries with group and group_level
Prior to this change `group` and `group_level` where overriding
each other. `group` was parsed as `group_level=exact`. This means
that the value of `group_level` was determined by the ordering
when both `group` and `group_level` were given.

This leads to confusion. With this change the following error will
be thrown in case both `group` and a `group_level` are specified
in the query:
Query parameter `group_level` is "not compatible with `group`

Change-Id: Idfa08e49760b159843ec49f484cdc063dd010c89
Reviewed-on: http://review.couchbase.org/31251
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>