MB-9889: Use PLAIN SASL to select bucket
With SASL authentication it's possible to select a bucket
where all subsequent UPR requests should apply to. This makes
it possible to use a non default bucket.

As there's one couch_upr instance per set view group we can
just supply the bucket the set view group operates on. It will
always be the same and never change through the group's lifetime.

Currently only PLAIN SASL without a password is implemented to
make it work. Later proper authentication should be implemented.
This means it currently works only on buckets that don't have a
password set.

Change-Id: I3adb025bd5fd502b12964334fe2f920682bd5745
Reviewed-on: http://review.couchbase.org/32327
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>