MB-10090 Fix receive_group_header buffered stdout reader method
receive_group_header method simulates blocked read of bytes in
given length by operating on the port interface in stream mode.
In stream mode, data is received as messages asynchronously.
If data accumulator received by this method is less than the
given length, it has to wait for data to arrive through port
and append to the existing data to extend the buffered data.

The current buffer was getting replaced by newly arrived data
rather than appending it and hence it indefinitely waits for
enough bytes to arrive. The test needs to use 1024 vbuckets
to go through this code path.

Change-Id: I757ecda73ac29ffba983aed64362cdf9b2f9914c
Reviewed-on: http://review.couchbase.org/33073
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>