MB-9974: Add sequence number and partition ID to #doc{} record.
In the UPR code you often need the sequence number and the partition ID
of a document. Currently it's supplied as a 3-tuple. With this change
the sequence number and the partition ID are directly moved into the

This is a preparation for a subsequent commit which will introduce proper
handling of multiple snapshots within one UPR stream request.

Change-Id: I704a5878dfa7b791861fab34cf29ef82a45f13c9
Reviewed-on: http://review.couchbase.org/33748
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>