MB-10472: Add support for requesting persisted items only
It is possible to make a stream request that includes the persisted items
only. They are all in a single snapshot.

In the fake UPR server you can set a function to simulate that behaviour.
The function takes the current high sequence and returns the sequence
number up to which the persisted-items-only snapshot should go. For now
it is set to return the whole data up to the high sequence.

Change-Id: I005732dc88dd387e5494a12b0b05cc6f42463047
Reviewed-on: http://review.couchbase.org/35904
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>