MB-10472: Build initial index with persisted items only
An UPR stream can consist of several snapshots. Only within snapshots
there's there guarantee that there's at most one mutation per document.

The persisted items are within a single snapshot. Hence we can use this
(potentially large) snapshot for initial index building.

This commit changes the updater logic to first request the persisted
items for every partition only and do the initial index build. Then the
updater is restarted to index the in-memory items. In case there are no
in-memory items, the updater isn't restarted. The in-memory items are
processed in normal incremental update fashion.

Change-Id: I9ae3e78b3afa25ebf57cf773ab048ab0a84f9abb
Reviewed-on: http://review.couchbase.org/35906
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>