MB-11023: Disable some collation tests
Some tests fail on Erlang 17.0. The reason is that Erlang 17.0
handles utf-8 binaries correctly and earlier releases did not.

This commit updates the tests to do the right thing. This makes
them fail under Erlang < 17.0. Hence they are commented out until
Couchbase bundles Erlang >= 17.0.

The issue only appears in escripts, this means that a compiled
module and the Erlang shell does not have the issue. This also
means that Couchbase on Erlang R1404 does the right thing as it
relies on compiled modules and not on escripts.

Change-Id: I62288e1afd0932d974f901127c7a203e85d6eacb
Reviewed-on: http://review.couchbase.org/37180
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: buildbot <build@couchbase.com>