Remove fragile assertions from btree tests
In the btree tests, it was tested whether a single key-value
insertion/deletion leads to an increase/decrease of the data size.
This is fragile as we compress the data. Hence even when you add
a new item, it could lead to a smaller total data size, due to better
compression.

This commit replaces those checks with a check of the file size
after all the values were inserted. The deletion test does already
do a similar check, hence no additional one was added.

Change-Id: I38103eadb47445f21196d31581e7ef79ec44a773
Reviewed-on: http://review.couchbase.org/37184
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>