MB-11472 Avoid btree update call when there is nothing to flush
The updater writer process data in batches of certain size.
If all kvs data is processed in last batch and current batch has
empty kvs, this is the final batch. We do not need to execute
btree update codepath for just updating the partition seqnos.

Since kvs list is empty and there is no active flush session,
id_btree sortfile will be nil and that case is currently unhandled
in the btree update code.

Change-Id: I9b18faf60c215560fb12c96909e83e02717c2216
Reviewed-on: http://review.couchbase.org/38489
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>