MB-11504 Remove stale delta logs from compactor delta log list
The index compactor can be started anytime without bothering if
an updater is already running or not. But updater always send its
log files to group coordinator for compactor's usage. But for a
compactor, it is only interested in delta files which are generated
after it is started. We should be able to discard any files created
before it is started. If a compactor is started while an updater is
running, it will ask for an explicit ack from updater. This ack can
be used to flush all old delta log files.

Stale files were being added to the compactor delta log list and
updater deletes those files from disk if a compactor is currently not
running. But log files remain in the compactor delta log list. When a
compactor is started next time, it tries to apply log files which are
not present on disk. This was leading failures in couch_view_file_merger.

This commit might also might slow down fragmentation by eliminating
stale delta log files being applied to the compacted index file.

Change-Id: Ic100a28337edcd3156cc011fc80b5225dae5e685
Reviewed-on: http://review.couchbase.org/38645
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>