MB-11486 Fix data corruption caused by commit d7d99
Fix for MB-11472 had a race condition that could cause data corruption.
The updater finish its operation when there is no id sortfile present
as well as kvs to be flushed is empty. But there could be a native updater
instance running in the background processing the last batch. We should
check and wait for currently running updater to finish before declaring
update operation as finished.

Change-Id: I404da8fbe77f1d28cbf6f5058e563712e0dee455
Reviewed-on: http://review.couchbase.org/38648
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>