Fix a race condition in unit tests
The tests 20-debug-params.t and 14-duplicated-keys-per-doc.t are not
waiting for all running updater processes to be finished before
shutdown of the view_group. This results in crash if the background
updater is trying to read paritition db files while shutdown group
command is in the process of deleting db files.

Change-Id: Ide04ea61b2d05a77261fcb1a3e1be68d2ce997d4
Reviewed-on: http://review.couchbase.org/38789
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Tested-by: Sarath Lakshman <sarath@couchbase.com>