MB-11563 Reset partition versions of recent partitions after rollback
When an index rollback is performed, we restore an old index header. In an
old index header, the partitions membership of the current index might be
different. Some recently added partitions may not be present in the index
snapshot pointed by the rollbacked header. Hence, we need to add entries
for newly added partitions which are not present in the partition lists
of old snapshot. We are currently updating seqnos list and cleanup bitmask.
But parititions version list also has to be updated. Otherwise, when the
updater tries to start an upr stream, it will not find the partition version
for a newly added partition and it would crash.

Change-Id: I4296cade0ff58929e4282e02ec6e154030f6205b
Reviewed-on: http://review.couchbase.org/38878
Tested-by: Nimish Gupta <nimish@couchbase.com>
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>