MB-11598 Update partition versions list correctly after compaction
The compactor delta log files list currently does not have partition
versions information. Hence, if the compactor and updater is running in
parallel and additional passive partitions are added, it could end up
in a situation where additional passive partitions information is not
updated in the partition versions list.

Fixed compactor to include partition versions list and update
them accordingly.

Change-Id: I79b8c10f4db389f5762f62e6033c28902ef04474
Reviewed-on: http://review.couchbase.org/39015
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Sarath Lakshman <sarath@couchbase.com>