Fix unit tests logic of trigger and wait for initial build
The initial build is performed in two phases. First phase
stream changes from on-disk snapshot and build the btree.
Updater is restarted to process in-memory snapshots (if any)
after the updater run for on-disk snapshot. Since the updater
is restarted during initial build, the unit test's mechanism of
waiting for initial updater's pid monitor message won't work.
The updater waiter will receive a DOWN message soon after first
phase is completed, but initial indexing may not have completed.

Hence, right way to wait for initial indexing to complete would
be to obtain the group with stale=false paramter.

Change-Id: I13f1ea081b0de57dc0d48fb0e178bb5711147332
Reviewed-on: http://review.couchbase.org/39300
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>