Fix rollback test to wait for updater exit only if pid is not nil
In certain cases, if the indexed seqnos is equal to current
partition seqnos, start_updater call from unit test will not
trigger the updater since there is no change remaining to be
applied. In that case, gen_server:call for start_updater will
return nil as pid. The updater waiter should not monitor and
wait in that case.

Change-Id: I9754ab93e135032dd483dafcc8b0d5d91dbd9416
Reviewed-on: http://review.couchbase.org/39403
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>