MB-11728 Fix upr fake server implementation of vbucket-seqno stats
When partition number is not specified in vbucket-seqno stats
request, ep-engine returns seqnos for all the existing partitions.
It does not return error response at all. Error response is only
returned when partition number is specified in the request.

Fixed fake server to not  return error response when vbucket-seqno
request is used.

Change-Id: Ifb72374f572cb615fbb50e58f473beef75a908e8
Reviewed-on: http://review.couchbase.org/39404
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>