Add per bucket auth to UPR client
The view engine uses a special case of authentication for UPR which
uses a single username/password combination for the whole cluster
and not per bucket. The desired bucket is then selected with the
select bucket command.

If the couch_upr_client is used outside of the view engine it is
desirable to be able to use the per bucket SASL authentication as
well. When used, the select bucket command returns "not implemented"
as the bucket is already selected through the username that was used
for authentication.

Change-Id: I0809de393b8ed11f9d54e74274ca077c05048d47
Reviewed-on: http://review.couchbase.org/39530
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>