MB-11728 Implement partition seqnos lazy cache
For stale=update_after queries, we are not really interested in
latest partition seqnos. Hence, we could cache the seqnos and make
use of it without paying the latency cost for stale=update_after queries.
The cache can be updated lazily based on ttl value (300 ms). Everytime
when a query obtains the cached seqnos, it checks whether the seqnos are
older than ttl value. If cached value is older than ttl, it will update
the cache by obtaining latest seqnos.

Change-Id: Iaea385692c8b948604e779df0f4a74942b6a4414
Reviewed-on: http://review.couchbase.org/39725
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>