External file sorting is no longer needed
The initial index build for the spatial views is now done completely
in C (Couchstore).

The incremental updates are still done in Erlang, though the ID b-tree
is already sorted natively in C. The native incremental updater itself,
has an option to do the sorting of the ID b-tree, hence the additional
external sorting step is no longer needed.

Change-Id: Ie639763a59fadb31c48fc4a0ec3ab66d87918a18
Reviewed-on: http://review.couchbase.org/39640
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>