MB-11838 Rename upr to dcp in couchdb.
Modify file names, modules and macros to reflect
the change in naming convention from upr to dcp.
Ensure macros are fully in upper case, Fix a few
that were in comments and others that were a mix
of upper and lower case. Updated .gitignore to
reflect change in directory names.

Change-Id: I9f464eef69f33c796686aae1b63352f764064b58
Reviewed-on: http://review.couchbase.org/40118
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Nimish Gupta <nimish@couchbase.com>
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Tested-by: Harsha H S <hhs.couchbase@gmail.com>