MB-11831 Maximum size of document that can be indexed is set to 20MB
Couchbase server restricts max size for JSON documents to 20 MB.
The view engine enforced a limit of 1 MB on documents that can
be indexed. This limit is now increased to 20 MB to ensure every
document gets indexed and not silently dropped by the view engine
if size of document exceeded previously enforced limit.

Change-Id: I4f357d003cc4de467c9facdf57e9c6f7d8bf55cf
Reviewed-on: http://review.couchbase.org/40175
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>