MB-11843 dcp_client: Fix reconnection handling with stream events
When the tcp connection used by dcp client is closed by ep-engine, the
dcp client tries to reestablish the tcp connection and dcp streams
which were active just before tcp connection was closed. Reestablishing
dcp connections and performing the repair is a multi-step operation and
complex. It is not possible to perform such a repair without considering
many cases such as change in failover log. Hence, it is cleaner to propogate
the connection error to the caller and the caller handle it gracefully.

The index updater is the consumer for dcp stream and it can retry the
index operation whenever dcp connection is closed rather than trying
repair can cause fatal errors.

Change-Id: Ieef7f527a37ab115d6a4c9c28af258915d4d0bc4
Reviewed-on: http://review.couchbase.org/40252
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Sarath Lakshman <sarath@couchbase.com>