Update indexed seqnos list correctly for incremental update
EP-Engine may send more mutations than requested endseqnos through DCP
when a stream is requested. We process all the mutations that arrived
from the DCP stream and update the index. But, the metadata list that
maintains the last indexed seqnos is incorrectly updated to the
requested endseqno for each partitions rather than seqno that we
received from the DCP stream. This could result in view engine
to request same data multiple times and process it redundantly.

Change-Id: I8371eda428f39fccf30948854a9b82c54f14a7ae
Reviewed-on: http://review.couchbase.org/40574
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Reviewed-by: Nimish Gupta <nimish@couchbase.com>