MB-12048: Make 3.0 mapreduce view group signature backwards compatible
The view engine uses a list of view records (#set_view{}) to calculate
the signature of a view group. Whenever this record is modified, the
signature will change.

In 3.0 the #set_view{} record changed in order to accommodate the
spatial views as well. This change lead to a different group signature
even if the design document didn't change. This rendered the views
created by Couchbase 2.x unusable. Those views would've been removed
and the indexing would've been started from scratch.

With this change the view groups in Couchbase 3.0 will have the same
signatures as they would have in 2.x.

NOTE: With this change the signatures of view groups change, this
means that all indexes created with a Couchbase 3.0 prior to this
commit will be rebuilt.

Change-Id: I901f6044cde5848eaaf18f63234e339cf3c80c82
Reviewed-on: http://review.couchbase.org/40915
Reviewed-by: Sriram Melkote <siri@couchbase.com>
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>