MB-12112 Add option to add search paths to unit test run script
It's now possible to add additional search paths for beam files via
the "-f" option. It expects a format that can be used in ERL_FLAGS.

This makes it possible for geocouch to add additional modules needed
to run the unit tests. It's about test helpers residing in the the
`test` subdirectory of the geocouch apps.

Change-Id: I339ff146a4920db0d59a8eafb02cf910e2dc02a1
Reviewed-on: http://review.couchbase.org/41762
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>